Move some test helpers from docs_test to fish_test
This is in preparation to make docs optional.

Backported from commit aabfea87c811de4c6bf6dcd01f64309a8864fead
